pyproject.toml
[project]
name = "scraping-work-schedule"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "flask>=3.0.3",
    "google-api-python-client>=2.158.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "python-dotenv>=1.0.1",
    "selenium>=4.27.1",
]
